+++ core/utils/events.py
@@ -89,7 +89,7 @@ async def on_member_join(self, member: discord.Member):
         if member.guild.id == 738425418637639775:
             join = langs.gts(member.joined_at, "en_gb", seconds=True)
             age = langs.td_dt(member.created_at, "en_gb")
-            await general.send(f"Welcome {member.name} to Senko Lair!\nJoin time: {join}\nAccount age: {age}", self.bot.get_channel(754425619336396851))
+            await general.send(f"Welcome {member.name} to Regaus' Playground!\nJoin time: {join}\nAccount age: {age}", self.bot.get_channel(754425619336396851))
     if member.guild.id in [568148147457490954, 738425418637639775] and member.id not in [302851022790066185]:
         if member.name[0] < "A":
             await member.edit(reason="De-hoist", nick=f"z{member.name}")
@@ -103,7 +103,7 @@ async def on_member_remove(self, member: discord.Member):
             await general.send(f"{member.name} just abandoned Senko Lair after surviving for {survival}...", self.bot.get_channel(610836120321785869))
         if member.guild.id == 738425418637639775:
             survival = langs.td_dt(member.joined_at, "en_gb")
-            await general.send(f"{member.name} just abandoned Senko Lair after surviving for {survival}...", self.bot.get_channel(754425619336396851))
+            await general.send(f"{member.name} just abandoned Regaus' Playground after surviving for {survival}...", self.bot.get_channel(754425619336396851))
         uid, gid = member.id, member.guild.id
         self.db.execute("DELETE FROM economy WHERE uid=? AND gid=?", (uid, gid))
         sel = self.db.fetchrow("SELECT * FROM leveling WHERE uid=? AND gid=?", (uid, gid))
@@ -118,10 +118,20 @@ async def on_member_remove(self, member: discord.Member):
 
 async def on_member_ban(self, guild: discord.Guild, user: discord.User or discord.Member):
     logger.log(self.bot.name, "members", f"{time.time()} > {self.bot.local_config['name']} > {user} ({user.id}) just got banned from {guild.name}")
+    message = f"{user} ({user.id}) has been **banned** from {guild.name}"
+    if guild.id == 568148147457490954:
+        await general.send(message, self.bot.get_channel(626028890451869707))
+    if guild.id == 738425418637639775:
+        await general.send(message, self.bot.get_channel(764469594303234078))
 
 
 async def on_member_unban(self, guild: discord.Guild, user: discord.User):
     logger.log(self.bot.name, "members", f"{time.time()} > {self.bot.local_config['name']} > {user} ({user.id}) just got unbanned from {guild.name}")
+    message = f"{user} ({user.id}) has been **unbanned** from {guild.name}"
+    if guild.id == 568148147457490954:
+        await general.send(message, self.bot.get_channel(626028890451869707))
+    if guild.id == 738425418637639775:
+        await general.send(message, self.bot.get_channel(764469594303234078))
 
 
 async def on_ready(self):